[ISSUE #10269]♻️Separate Proxy ingress and Store logical backend reads
