+++ rocketmq-proxy-core/Cargo.toml
@@ -27,13 +27,13 @@ rust-version.workspace = true
 
 [features]
 default = []
+grpc-bindings = ["dep:tonic", "dep:tonic-prost"]
 
 [dependencies]
 rocketmq-error = { workspace = true }
 rocketmq-model = { workspace = true }
 rocketmq-protocol = { workspace = true }
 rocketmq-runtime = { workspace = true }
-rocketmq-transport = { workspace = true, default-features = false }
 
 dashmap.workspace = true
 serde.workspace = true
@@ -45,8 +45,8 @@ bytes.workspace = true
 
 prost = "0.14"
 prost-types = "0.14"
-tonic = { version = "0.14.6", features = ["transport", "codegen"] }
-tonic-prost = "0.14"
+tonic = { version = "0.14.6", optional = true, features = ["transport", "codegen"] }
+tonic-prost = { version = "0.14", optional = true }
 
 [build-dependencies]
 tonic-prost-build = "0.14.6"
