+++ rocketmq-proxy-core/src/status.rs
@@ -13,14 +13,11 @@
 // limitations under the License.
 
 use rocketmq_error::GrpcPayloadCode;
-use rocketmq_error::GrpcStatusCode;
 use rocketmq_error::PROXY_METADATA_INVALID;
 use rocketmq_error::PROXY_SETTINGS_UNAVAILABLE;
 use rocketmq_error::PROXY_TRANSPORT_UNAVAILABLE;
 use rocketmq_model::result::SendResult;
 use rocketmq_model::result::SendStatus;
-use tonic::Code as TonicCode;
-use tonic::Status as TonicStatus;
 
 use crate::error::ProxyError;
 use crate::proto::v2;
@@ -112,15 +109,6 @@ impl ProxyStatusMapper {
         Self::from_error_payload(error).into()
     }
 
-    pub fn to_tonic_status(error: &ProxyError) -> TonicStatus {
-        let descriptor = error.descriptor();
-        let grpc = descriptor.projection().grpc();
-        TonicStatus::new(
-            Self::grpc_status_to_tonic_code(grpc.status),
-            descriptor.public_message(),
-        )
-    }
-
     fn grpc_payload_to_code(payload: GrpcPayloadCode) -> v2::Code {
         match payload {
             GrpcPayloadCode::InternalError => v2::Code::InternalError,
@@ -153,452 +141,5 @@ impl ProxyStatusMapper {
             GrpcPayloadCode::LiteSubscriptionQuotaExceeded => v2::Code::LiteSubscriptionQuotaExceeded,
             GrpcPayloadCode::MessagePropertyConflictWithType => v2::Code::MessagePropertyConflictWithType,
         }
-    }
-
-    fn grpc_status_to_tonic_code(status: GrpcStatusCode) -> TonicCode {
-        match status {
-            GrpcStatusCode::InvalidArgument => TonicCode::InvalidArgument,
-            GrpcStatusCode::Unauthenticated => TonicCode::Unauthenticated,
-            GrpcStatusCode::PermissionDenied => TonicCode::PermissionDenied,
-            GrpcStatusCode::NotFound => TonicCode::NotFound,
-            GrpcStatusCode::DeadlineExceeded => TonicCode::DeadlineExceeded,
-            GrpcStatusCode::ResourceExhausted => TonicCode::ResourceExhausted,
-            GrpcStatusCode::FailedPrecondition => TonicCode::FailedPrecondition,
-            GrpcStatusCode::AlreadyExists => TonicCode::AlreadyExists,
-            GrpcStatusCode::Aborted => TonicCode::Aborted,
-            GrpcStatusCode::Unimplemented => TonicCode::Unimplemented,
-            GrpcStatusCode::Unavailable => TonicCode::Unavailable,
-            GrpcStatusCode::DataLoss => TonicCode::DataLoss,
-            GrpcStatusCode::Internal => TonicCode::Internal,
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use rocketmq_error::GrpcStatusCode;
-    use rocketmq_error::PublicErrorView;
-    use rocketmq_protocol::code::response_code::ResponseCode;
-
-    use super::ProxyStatusMapper;
-    use crate::error::canonical;
-    use crate::error::ProxyError;
-    use crate::proto::v2;
-
-    #[test]
-    fn client_id_required_maps_to_expected_code() {
-        let status = ProxyStatusMapper::from_error(&ProxyError::ClientIdRequired);
-        assert_eq!(status.code, v2::Code::ClientIdRequired as i32);
-    }
-
-    #[test]
-    fn route_not_found_maps_to_topic_not_found() {
-        let status = ProxyStatusMapper::from_error(&ProxyError::Canonical(canonical::route_not_found("TestTopic")));
-        assert_eq!(status.code, v2::Code::TopicNotFound as i32);
-    }
-
-    #[test]
-    fn route_not_found_requires_typed_error_instead_of_display_text() {
-        let status = ProxyStatusMapper::from_error(&ProxyError::Canonical(canonical::argument(
-            "CODE: 17  DESC: No topic route info in name server for the topic: TestTopic",
-        )));
-
-        assert_eq!(status.code, v2::Code::BadRequest as i32);
-        assert_eq!(
-            ProxyStatusMapper::to_tonic_status(&ProxyError::Canonical(canonical::argument(
-                "CODE: 17  DESC: No topic route info in name server for the topic: TestTopic",
-            )))
-            .code(),
-            tonic::Code::InvalidArgument
-        );
-    }
-
-    #[test]
-    fn invalid_metadata_prefers_tonic_status() {
-        let error = ProxyError::invalid_metadata("grpc-timeout must be valid");
-        assert!(ProxyStatusMapper::should_use_tonic_status(&error));
-        assert_eq!(
-            ProxyStatusMapper::to_tonic_status(&error).code(),
-            tonic::Code::InvalidArgument
-        );
-    }
-
-    #[test]
-    fn transport_error_maps_to_unavailable_tonic_status() {
-        let error = ProxyError::Transport {
-            message: "cluster worker unavailable".to_owned(),
-        };
-        assert!(ProxyStatusMapper::should_use_tonic_status(&error));
-        assert_eq!(
-            ProxyStatusMapper::to_tonic_status(&error).code(),
-            tonic::Code::Unavailable
-        );
-    }
-
-    #[test]
-    fn phase7_broker_auth_and_permission_response_codes_map_to_grpc_payload_codes() {
-        for (response_code, expected_grpc_code) in [
-            (ResponseCode::NoPermission, v2::Code::Forbidden),
-            (ResponseCode::UserNotExist, v2::Code::NotFound),
-            (ResponseCode::PolicyNotExist, v2::Code::NotFound),
-        ] {
-            let error = ProxyError::BrokerResponse(canonical::broker_response(
-                "AUTH_ADMIN",
-                response_code.to_i32(),
-                None,
-                "auth failed",
-            ));
-            let status = ProxyStatusMapper::from_error(&error);
-            assert_eq!(status.code, expected_grpc_code as i32);
-        }
-
-        let authentication_error =
-            ProxyError::Canonical(canonical::authentication_failed("authenticate", "bad credentials"));
-        let status = ProxyStatusMapper::from_error(&authentication_error);
-        assert_eq!(status.code, v2::Code::Unauthorized as i32);
-    }
-
-    #[test]
-    fn phase7_request_code_not_supported_maps_to_unsupported_and_unimplemented_transport() {
-        let error = ProxyError::BrokerResponse(canonical::broker_response(
-            "REMOTING",
-            ResponseCode::RequestCodeNotSupported.to_i32(),
-            None,
-            "request code not supported",
-        ));
-
-        let payload_status = ProxyStatusMapper::from_error(&error);
-        assert_eq!(payload_status.code, v2::Code::Unsupported as i32);
-        assert_eq!(
-            ProxyStatusMapper::to_tonic_status(&error).code(),
-            tonic::Code::Unimplemented
-        );
-    }
-
-    #[test]
-    fn normalized_broker_responses_use_descriptor_payload_and_tonic_status() {
-        for (response_code, expected_grpc_code, expected_tonic_code) in [
-            (
-                ResponseCode::NoPermission,
-                v2::Code::Forbidden,
-                tonic::Code::PermissionDenied,
-            ),
-            (
-                ResponseCode::TopicNotExist,
-                v2::Code::TopicNotFound,
-                tonic::Code::NotFound,
-            ),
-            (
-                ResponseCode::SubscriptionGroupNotExist,
-                v2::Code::ConsumerGroupNotFound,
-                tonic::Code::NotFound,
-            ),
-            (ResponseCode::UserNotExist, v2::Code::NotFound, tonic::Code::NotFound),
-            (ResponseCode::PolicyNotExist, v2::Code::NotFound, tonic::Code::NotFound),
-            (
-                ResponseCode::QueryNotFound,
-                v2::Code::OffsetNotFound,
-                tonic::Code::NotFound,
-            ),
-            (
-                ResponseCode::PullOffsetMoved,
-                v2::Code::IllegalOffset,
-                tonic::Code::InvalidArgument,
-            ),
-            (
-                ResponseCode::RequestCodeNotSupported,
-                v2::Code::Unsupported,
-                tonic::Code::Unimplemented,
-            ),
-            (ResponseCode::SystemBusy, v2::Code::InternalError, tonic::Code::Internal),
-        ] {
-            let error = ProxyError::BrokerResponse(canonical::broker_response(
-                "BROKER",
-                response_code.to_i32(),
-                None,
-                "secret broker response\r\nC:\\private\\broker.conf",
-            ));
-            let payload_status = ProxyStatusMapper::from_error(&error);
-            assert_eq!(payload_status.code, expected_grpc_code as i32);
-            assert_eq!(ProxyStatusMapper::to_tonic_status(&error).code(), expected_tonic_code);
-            assert!(!payload_status.message.contains("secret"));
-            assert!(!payload_status.message.contains("private"));
-        }
-    }
-
-    #[test]
-    fn auth_config_errors_map_to_bad_request_payload_and_transport_status() {
-        let error = ProxyError::Canonical(canonical::configuration_invalid(
-            "auth.authorization",
-            "provider not ready",
-        ));
-        let payload_status = ProxyStatusMapper::from_error(&error);
-        assert_eq!(payload_status.code, v2::Code::BadRequest as i32);
-        assert_eq!(
-            ProxyStatusMapper::to_tonic_status(&error).code(),
-            tonic::Code::InvalidArgument
-        );
-    }
-
-    #[test]
-    fn canonical_errors_use_catalog_grpc_boundary_spec() {
-        let retry_exhausted = ProxyError::Canonical(canonical::retry_budget_exhausted("GID_test", 3, 3));
-        let retry_payload = ProxyStatusMapper::from_error(&retry_exhausted);
-        assert_eq!(retry_payload.code, v2::Code::TooManyRequests as i32);
-        assert_eq!(
-            ProxyStatusMapper::to_tonic_status(&retry_exhausted).code(),
-            tonic::Code::ResourceExhausted
-        );
-
-        let not_master = ProxyError::Canonical(canonical::not_master("127.0.0.1:10911"));
-        let not_master_payload = ProxyStatusMapper::from_error(&not_master);
-        assert_eq!(not_master_payload.code, v2::Code::InternalError as i32);
-        assert_eq!(
-            ProxyStatusMapper::to_tonic_status(&not_master).code(),
-            tonic::Code::FailedPrecondition
-        );
-    }
-
-    #[test]
-    fn data_loss_status_maps_to_tonic_data_loss() {
-        assert_eq!(
-            ProxyStatusMapper::grpc_status_to_tonic_code(GrpcStatusCode::DataLoss),
-            tonic::Code::DataLoss
-        );
-    }
-
-    #[test]
-    fn newly_declared_statuses_map_exhaustively_to_tonic() {
-        for (status, expected) in [
-            (GrpcStatusCode::AlreadyExists, tonic::Code::AlreadyExists),
-            (GrpcStatusCode::Aborted, tonic::Code::Aborted),
-        ] {
-            assert_eq!(ProxyStatusMapper::grpc_status_to_tonic_code(status), expected);
-        }
-    }
-
-    #[test]
-    fn canonical_error_payload_message_uses_public_message() {
-        let error = ProxyError::Canonical(canonical::configuration_invalid(
-            "auth.authorization",
-            "provider not ready",
-        ));
-        let expected_message = error.descriptor().public_message().to_owned();
-
-        let status = ProxyStatusMapper::from_error(&error);
-
-        assert_eq!(status.message, expected_message);
-    }
-
-    #[test]
-    fn all_local_proxy_errors_use_fixed_catalog_output_and_preserve_wire_codes() {
-        const MALICIOUS: &str = "secret-token\r\nC:\\private\\proxy.pem";
-
-        let cases = vec![
-            (
-                ProxyError::ClientIdRequired,
-                "proxy.client.id.required",
-                v2::Code::ClientIdRequired,
-                tonic::Code::InvalidArgument,
-                1,
-            ),
-            (
-                ProxyError::UnrecognizedClientType(31337),
-                "proxy.client.type.unrecognized",
-                v2::Code::UnrecognizedClientType,
-                tonic::Code::InvalidArgument,
-                1,
-            ),
-            (
-                ProxyError::not_implemented(MALICIOUS),
-                "proxy.capability.unsupported",
-                v2::Code::NotImplemented,
-                tonic::Code::Unimplemented,
-                3,
-            ),
-            (
-                ProxyError::too_many_requests(MALICIOUS),
-                "proxy.capacity.exhausted",
-                v2::Code::TooManyRequests,
-                tonic::Code::ResourceExhausted,
-                2,
-            ),
-            (
-                ProxyError::Draining,
-                "proxy.request.draining",
-                v2::Code::InternalError,
-                tonic::Code::Unavailable,
-                1,
-            ),
-            (
-                ProxyError::invalid_metadata(MALICIOUS),
-                "proxy.metadata.invalid",
-                v2::Code::BadRequest,
-                tonic::Code::InvalidArgument,
-                1,
-            ),
-            (
-                ProxyError::Transport {
-                    message: MALICIOUS.to_owned(),
-                },
-                "proxy.transport.unavailable",
-                v2::Code::InternalError,
-                tonic::Code::Unavailable,
-                1,
-            ),
-            (
-                ProxyError::illegal_message_id(MALICIOUS),
-                "proxy.message.id.invalid",
-                v2::Code::IllegalMessageId,
-                tonic::Code::InvalidArgument,
-                13,
-            ),
-            (
-                ProxyError::invalid_transaction_id(MALICIOUS),
-                "proxy.transaction.id.invalid",
-                v2::Code::InvalidTransactionId,
-                tonic::Code::InvalidArgument,
-                1,
-            ),
-            (
-                ProxyError::illegal_message_group(MALICIOUS),
-                "proxy.message.group.invalid",
-                v2::Code::IllegalMessageGroup,
-                tonic::Code::InvalidArgument,
-                13,
-            ),
-            (
-                ProxyError::illegal_delivery_time(MALICIOUS),
-                "proxy.delivery.time.invalid",
-                v2::Code::IllegalDeliveryTime,
-                tonic::Code::InvalidArgument,
-                13,
-            ),
-            (
-                ProxyError::illegal_polling_time(MALICIOUS),
-                "proxy.polling.time.invalid",
-                v2::Code::IllegalPollingTime,
-                tonic::Code::InvalidArgument,
-                1,
-            ),
-            (
-                ProxyError::illegal_offset(MALICIOUS),
-                "proxy.offset.invalid",
-                v2::Code::IllegalOffset,
-                tonic::Code::InvalidArgument,
-                21,
-            ),
-            (
-                ProxyError::illegal_invisible_time(MALICIOUS),
-                "proxy.invisible.time.invalid",
-                v2::Code::IllegalInvisibleTime,
-                tonic::Code::InvalidArgument,
-                1,
-            ),
-            (
-                ProxyError::illegal_filter_expression(MALICIOUS),
-                "proxy.filter.expression.invalid",
-                v2::Code::IllegalFilterExpression,
-                tonic::Code::InvalidArgument,
-                23,
-            ),
-            (
-                ProxyError::invalid_receipt_handle(MALICIOUS),
-                "proxy.receipt.handle.invalid",
-                v2::Code::InvalidReceiptHandle,
-                tonic::Code::InvalidArgument,
-                1,
-            ),
-            (
-                ProxyError::illegal_lite_topic(MALICIOUS),
-                "proxy.lite_topic.invalid",
-                v2::Code::IllegalLiteTopic,
-                tonic::Code::InvalidArgument,
-                1,
-            ),
-            (
-                ProxyError::lite_subscription_quota_exceeded(MALICIOUS),
-                "proxy.lite_subscription.quota_exceeded",
-                v2::Code::LiteSubscriptionQuotaExceeded,
-                tonic::Code::ResourceExhausted,
-                1,
-            ),
-            (
-                ProxyError::message_property_conflict(MALICIOUS),
-                "proxy.message.property_conflict",
-                v2::Code::MessagePropertyConflictWithType,
-                tonic::Code::InvalidArgument,
-                13,
-            ),
-            (
-                ProxyError::settings_unavailable(MALICIOUS),
-                "proxy.settings.unavailable",
-                v2::Code::InternalError,
-                tonic::Code::FailedPrecondition,
-                1,
-            ),
-        ];
-
-        for (error, expected_code, expected_payload, expected_tonic, expected_remoting) in cases {
-            assert_eq!(error.descriptor().code().as_str(), expected_code);
-            assert_eq!(
-                error.descriptor().projection().remoting().code.as_i32(),
-                expected_remoting,
-                "{expected_code}"
-            );
-
-            let context = error.context();
-            let public = PublicErrorView::try_new(error.descriptor(), &context)
-                .expect("Proxy local context must match its catalog descriptor");
-            assert_eq!(public.fields().count(), 0, "{expected_code}");
-
-            let payload = ProxyStatusMapper::from_error_payload(&error);
-            let tonic = ProxyStatusMapper::to_tonic_status(&error);
-            assert_eq!(payload.code(), expected_payload as i32, "{expected_code}");
-            assert_eq!(tonic.code(), expected_tonic, "{expected_code}");
-            assert_eq!(
-                payload.message(),
-                error.descriptor().public_message(),
-                "{expected_code}"
-            );
-            assert_eq!(tonic.message(), error.descriptor().public_message(), "{expected_code}");
-            for output in [payload.message(), tonic.message()] {
-                assert!(!output.contains("secret-token"), "{expected_code}");
-                assert!(!output.contains("private"), "{expected_code}");
-                assert!(!output.chars().any(char::is_control), "{expected_code}");
-            }
-            assert_eq!(
-                ProxyStatusMapper::should_use_tonic_status(&error),
-                matches!(
-                    expected_code,
-                    "proxy.metadata.invalid" | "proxy.transport.unavailable" | "proxy.settings.unavailable"
-                ),
-                "{expected_code}"
-            );
-        }
-    }
-
-    #[test]
-    fn local_proxy_errors_use_catalog_descriptor_mapping() {
-        let lite_topic = ProxyError::illegal_lite_topic("not an LMQ");
-        assert_eq!(lite_topic.descriptor().code().as_str(), "proxy.lite_topic.invalid");
-        let lite_topic_status = ProxyStatusMapper::from_error(&lite_topic);
-        assert_eq!(lite_topic_status.code, v2::Code::IllegalLiteTopic as i32);
-        assert_eq!(
-            ProxyStatusMapper::to_tonic_status(&lite_topic).code(),
-            tonic::Code::InvalidArgument
-        );
-
-        let quota = ProxyError::lite_subscription_quota_exceeded("subscription limit reached");
-        assert_eq!(
-            quota.descriptor().code().as_str(),
-            "proxy.lite_subscription.quota_exceeded"
-        );
-        let quota_status = ProxyStatusMapper::from_error(&quota);
-        assert_eq!(quota_status.code, v2::Code::LiteSubscriptionQuotaExceeded as i32);
-        assert_eq!(
-            ProxyStatusMapper::to_tonic_status(&quota).code(),
-            tonic::Code::ResourceExhausted
-        );
     }
 }
