+++ rocketmq-proxy/Cargo.toml
@@ -49,7 +49,7 @@ tls = ["dep:arc-swap", "dep:parking_lot", "dep:tokio-rustls", "rocketmq-transpor
 rocketmq-error       = { workspace = true }
 rocketmq-model       = { workspace = true }
 rocketmq-protocol    = { workspace = true }
-rocketmq-proxy-core  = { workspace = true }
+rocketmq-proxy-core  = { workspace = true, features = ["grpc-bindings"] }
 rocketmq-proxy-cluster = { workspace = true, optional = true }
 rocketmq-proxy-local = { workspace = true, optional = true }
 rocketmq-auth        = { workspace = true, features = ["grpc"] }
@@ -86,6 +86,7 @@ parking_lot = { workspace = true, optional = true }
 tokio-rustls = { version = "0.26", optional = true }
 
 [dev-dependencies]
+prost = "0.14"
 rocketmq-transport = { workspace = true, features = ["test-support"] }
 tokio = { workspace = true, features = ["test-util"] }
 criterion = { version = "0.8", features = ["html_reports"] }
