+++ rocketmq-proxy/src/config.rs
@@ -30,7 +30,8 @@ pub use rocketmq_proxy_core::config::GrpcConfig;
 pub use rocketmq_proxy_core::config::GrpcTlsClientAuth;
 pub use rocketmq_proxy_core::config::GrpcTlsConfig;
 pub use rocketmq_proxy_core::config::ProxyMode;
-pub use rocketmq_proxy_core::config::RemotingConfig;
+mod remoting;
+pub use remoting::RemotingConfig;
 pub use rocketmq_proxy_core::config::RuntimeConfig;
 pub use rocketmq_proxy_core::config::SessionConfig;
 pub use rocketmq_proxy_core::SettingsConfig;
