+++ rocketmq-proxy/src/ingress.rs
@@ -12,12 +12,5 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::ProxyMessage;
-
-// Producer gRPC ingress planning.
-
-const TRANSACTION_PREPARED_PROPERTY: &str = "TRAN_MSG";
-
-pub fn is_transaction_message(message: &ProxyMessage) -> bool {
-    message.property(TRANSACTION_PREPARED_PROPERTY).is_some()
-}
+//! Network ingress orchestration over the neutral Proxy contracts.
+pub mod grpc;
