+++ rocketmq-proxy/src/ingress/grpc/service/topic.rs
@@ -13,11 +13,11 @@
 // limitations under the License.
 
 use crate::proto::v2;
-use crate::ProxyContextWithPrincipal;
+use rocketmq_proxy_core::ProxyContextWithPrincipal;
 // Topic gRPC ingress validation.
 
-use crate::ProxyError;
-use crate::ProxyResult;
+use rocketmq_proxy_core::ProxyError;
+use rocketmq_proxy_core::ProxyResult;
 
 pub fn validate_client_context<P>(context: &ProxyContextWithPrincipal<P>) -> ProxyResult<&str> {
     context.require_client_id()
