+++ rocketmq-store/src/capability.rs
@@ -23,6 +23,7 @@ mod read;
 pub use read::BrokerReadStore;
 pub use read::ConsumeQueueStatistics;
 
+use crate::base::backend_read_ops::BackendReadOps;
 use std::any::Any;
 use std::collections::HashMap;
 use std::collections::HashSet;
@@ -284,7 +285,7 @@ pub trait BrokerAdminStore: BrokerReplicationStore {
 
     /// Returns queue counts without exposing the backend's queue table or lock.
     fn consume_queue_statistics(&self) -> ConsumeQueueStatistics {
-        read::consume_queue_statistics(self.backend())
+        BackendReadOps::consume_queue_statistics(self.backend())
     }
 
     fn delete_topics(&self, delete_topics: Vec<&CheetahString>) -> i32 {
@@ -614,7 +615,7 @@ impl<MS: BackendOps> MessageStoreReadPort for MS {
         offset: i64,
         max_messages: i32,
     ) -> Option<GetMessageResult> {
-        BackendOps::get_message(self, group, topic, queue_id, offset, max_messages, None).await
+        BackendReadOps::get_message(self, group, topic, queue_id, offset, max_messages, None).await
     }
 
     async fn get_message_with_size_limit(
@@ -626,7 +627,7 @@ impl<MS: BackendOps> MessageStoreReadPort for MS {
         max_messages: i32,
         max_total_size: i32,
     ) -> Option<GetMessageResult> {
-        BackendOps::get_message_with_size_limit(
+        BackendReadOps::get_message_with_size_limit(
             self,
             group,
             topic,
@@ -647,13 +648,13 @@ impl<MS: BackendOps> MessageStoreReadPort for MS {
         begin: i64,
         end: i64,
     ) -> Option<QueryMessageResult> {
-        BackendOps::query_message(self, topic, key, max_messages, begin, end).await
+        BackendReadOps::query_message(self, topic, key, max_messages, begin, end).await
     }
 
     fn select_message(&self, physical_offset: i64, size: Option<i32>) -> Option<SelectMappedBufferResult> {
         match size {
-            Some(size) => BackendOps::select_one_message_by_offset_with_size(self, physical_offset, size),
-            None => BackendOps::select_one_message_by_offset(self, physical_offset),
+            Some(size) => BackendReadOps::select_one_message_by_offset_with_size(self, physical_offset, size),
+            None => BackendReadOps::select_one_message_by_offset(self, physical_offset),
         }
     }
 }
