+++ rocketmq-store/tests/message_store/local_file_message_store/unit.rs
@@ -88,6 +88,7 @@ use super::ReputMessageService;
 use super::ReputMessageServiceInner;
 use super::StoreRootLeaseState;
 use crate::base::backend_ops::BackendOps;
+use crate::base::backend_read_ops::BackendReadOps;
 use crate::base::commit_log_dispatcher::CommitLogDispatcher;
 use crate::base::dispatch_request::DispatchRequest;
 use crate::base::message_arriving_listener::MessageArrivingListener;
